Revert "[Ahuna] Lib and routes to get assistants according to views (#2778)"
This reverts commit 0af26bc9858e0862b19f285050b23cc129819b7b.
