+++ front/components/assistant/conversation/AgentSuggestion.tsx
@@ -26,7 +26,6 @@ export function AgentSuggestion({
 }) {
   const { agentConfigurations } = useAgentConfigurations({
     workspaceId: owner.sId,
-    agentsGetViewType: { conversationId: conversation.sId },
   });
   const sendNotification = useContext(SendNotificationsContext);
