+++ front/components/assistant/conversation/InputBar.tsx
@@ -62,7 +62,6 @@ function AgentListImpl(
     visible,
     filter,
     position,
-    conversationId,
   }: {
     owner: WorkspaceType;
     visible: boolean;
@@ -71,7 +70,6 @@ function AgentListImpl(
       bottom: number;
       left: number;
     };
-    conversationId: string | null;
   },
   ref: ForwardedRef<{
     prev: () => void;
@@ -88,7 +86,6 @@ function AgentListImpl(
 
   const { agentConfigurations } = useAgentConfigurations({
     workspaceId: owner.sId,
-    agentsGetViewType: conversationId ? { conversationId } : "list",
   });
 
   const activeAgents = agentConfigurations.filter((a) => a.status === "active");
@@ -286,7 +283,6 @@ export function AssistantInputBar({
 
   const { agentConfigurations } = useAgentConfigurations({
     workspaceId: owner.sId,
-    agentsGetViewType: conversationId ? { conversationId } : "list",
   });
   const sendNotification = useContext(SendNotificationsContext);
 
@@ -477,7 +473,6 @@ export function AssistantInputBar({
         filter={agentListFilter}
         ref={agentListRef}
         position={agentListPosition}
-        conversationId={conversationId}
       />
 
       {generationContext.generatingMessageIds.length > 0 && (
