+++ front/lib/api/assistant/configuration.ts
@@ -1,11 +1,6 @@
-import {
-  AgentMention,
-  AgentRelationOverrideType,
-  SupportedModel,
-} from "@dust-tt/types";
+import { SupportedModel } from "@dust-tt/types";
 import { DustAppRunConfigurationType } from "@dust-tt/types";
 import {
-  AgentsGetViewType,
   DataSourceConfiguration,
   isTemplatedQuery,
   isTimeFrame,
@@ -20,7 +15,6 @@ import {
   AgentGenerationConfigurationType,
   AgentStatus,
 } from "@dust-tt/types";
-import { isSupportedModel } from "@dust-tt/types";
 import { Op, Transaction } from "sequelize";
 
 import {
@@ -28,6 +22,7 @@ import {
   getGlobalAgents,
   isGlobalAgentId,
 } from "@app/lib/api/assistant/global_agents";
+import { isSupportedModel } from "@app/lib/assistant";
 import { Authenticator } from "@app/lib/auth";
 import { front_sequelize } from "@app/lib/databases";
 import {
@@ -36,13 +31,9 @@ import {
   AgentDustAppRunConfiguration,
   AgentGenerationConfiguration,
   AgentRetrievalConfiguration,
-  Conversation,
   DataSource,
-  Mention,
-  Message,
   Workspace,
 } from "@app/lib/models";
-import { AgentUserRelation } from "@app/lib/models/assistant/agent";
 import { generateModelSId } from "@app/lib/utils";
 
 /**
@@ -83,14 +74,6 @@ export async function getAgentConfiguration(
       {
         model: AgentDustAppRunConfiguration,
         as: "dustAppRunConfiguration",
-      },
-      {
-        model: AgentUserRelation,
-        where: {
-          userId: auth.user()?.id,
-        },
-        attributes: ["relation"],
-        required: false,
       },
     ],
     limit: 1,
@@ -206,16 +189,11 @@ export async function getAgentConfiguration(
     };
   }
 
-  const relationOverride =
-    agent.relationOverrides?.length > 0
-      ? agent.relationOverrides[0].relation
-      : null;
   return {
     id: agent.id,
     sId: agent.sId,
     version: agent.version,
     scope: agent.scope,
-    relationOverride,
     name: agent.name,
     pictureUrl: agent.pictureUrl,
     description: agent.description,
@@ -226,162 +204,47 @@ export async function getAgentConfiguration(
 }
 
 /**
- * Get agent configurations for the workspace, optionally whose names
+ * Get the list agent configuration for the workspace, optionally whose names
  * match a prefix
- * @param agentsGetView the kind of list of agents we want to get, see AgentsGetViewType
  */
 export async function getAgentConfigurations(
   auth: Authenticator,
-  agentsGetView: AgentsGetViewType,
   agentPrefix?: string
 ): Promise<AgentConfigurationType[] | []> {
   const owner = auth.workspace();
   if (!owner) {
     throw new Error("Unexpected `auth` without `workspace`.");
   }
-  if (!auth.isUser()) {
-    throw new Error("Unexpected `auth` from outside workspace.");
-  }
 
-  const where = {
-    workspaceId: owner.id,
-    status: "active",
-    ...(agentPrefix ? { name: { [Op.iLike]: `${agentPrefix}%` } } : {}),
-  };
+  const rawAgents = await AgentConfiguration.findAll({
+    where: {
+      workspaceId: owner.id,
+      status: "active",
+      ...(agentPrefix
+        ? {
+            name: {
+              [Op.iLike]: `${agentPrefix}%`,
+            },
+          }
+        : {}),
+    },
+    order: [["name", "ASC"]],
+  });
+
+  const agents = (
+    await Promise.all(
+      rawAgents.map(async (a) => {
+        return await getAgentConfiguration(auth, a.sId);
+      })
+    )
+  ).filter((a) => a !== null) as AgentConfigurationType[];
 
   const globalAgents = (await getGlobalAgents(auth)).filter(
     (a) =>
       !agentPrefix || a.name.toLowerCase().startsWith(agentPrefix.toLowerCase())
   );
 
-  const getLocalAgentsFromIds = async (agentIds: string[]) => {
-    return (
-      await Promise.all(agentIds.map((a) => getAgentConfiguration(auth, a)))
-    ).filter((a) => a !== null) as AgentConfigurationType[];
-  };
-
-  if (agentsGetView === "super_user") {
-    if (!auth.isDustSuperUser()) {
-      throw new Error("superuser view is for dust superusers only.");
-    }
-    const agentIds = (
-      await AgentConfiguration.findAll({
-        where,
-        attributes: ["sId"],
-      })
-    ).map((a) => a.sId);
-    return [...(await getLocalAgentsFromIds(agentIds)), ...globalAgents];
-  }
-
-  if (agentsGetView === "all") {
-    const agentIds = (
-      await AgentConfiguration.findAll({
-        where: {
-          ...where,
-          scope: { [Op.in]: ["published", "workspace"] },
-        },
-        attributes: ["sId"],
-      })
-    ).map((a) => a.sId);
-    return [...(await getLocalAgentsFromIds(agentIds)), ...globalAgents];
-  }
-
-  if (agentsGetView === "list") {
-    const user = auth.user();
-    if (!user) {
-      throw new Error("List view is specific to a user.");
-    }
-    const agentIds = (
-      await AgentConfiguration.findAll({
-        where: {
-          ...where,
-          [Op.or]: [
-            { scope: { [Op.in]: ["published", "workspace"] } },
-            { authorId: user.id },
-          ],
-        },
-        attributes: ["sId", "scope", "name"],
-      })
-    ).map((a) => a.sId);
-    const localAgents = (await getLocalAgentsFromIds(agentIds)).filter((a) => {
-      if (a.scope === "workspace") {
-        return a.relationOverride !== "not-in-list";
-      }
-      if (a.scope === "published") {
-        return a.relationOverride === "in-list";
-      }
-      return true; // user's private agents should be returned
-    }) as AgentConfigurationType[];
-    return [...localAgents, ...globalAgents];
-  }
-
-  // conversation view
-  if (typeof agentsGetView === "object" && agentsGetView.conversationId) {
-    const user = auth.user();
-    if (!user) {
-      throw new Error("Conversation view is specific to a user.");
-    }
-    const [agents, mentions] = await Promise.all([
-      AgentConfiguration.findAll({
-        where: {
-          ...where,
-          [Op.or]: [
-            { scope: { [Op.in]: ["published", "workspace"] } },
-            { authorId: user && user.id },
-          ],
-        },
-        attributes: ["sId", "scope", "name"],
-      }),
-      getConversationMentions(agentsGetView.conversationId),
-    ]);
-    const agentIds = agents.map((a) => a.sId);
-    const mentionedAgentIds = mentions.map((m) => m.configurationId);
-    const localAgents = (await getLocalAgentsFromIds(agentIds)).filter((a) => {
-      if (mentionedAgentIds.includes(a.sId)) {
-        return true;
-      }
-      if (a.scope === "workspace") {
-        return a.relationOverride !== "not-in-list";
-      }
-      if (a.scope === "published") {
-        return a.relationOverride === "in-list";
-      }
-      return true; // user's private agents should be returned
-    }) as AgentConfigurationType[];
-    return [...localAgents, ...globalAgents];
-  }
-
-  throw new Error(`Unknown agentsGetView ${agentsGetView}`);
-}
-
-async function getConversationMentions(
-  conversationId: string
-): Promise<AgentMention[]> {
-  const mentions = await Mention.findAll({
-    attributes: ["agentConfigurationId"],
-    where: {
-      agentConfigurationId: {
-        [Op.ne]: null,
-      },
-    },
-    include: [
-      {
-        model: Message,
-        attributes: [],
-        include: [
-          {
-            model: Conversation,
-            as: "conversation",
-            attributes: [],
-            where: { sId: conversationId },
-          },
-        ],
-      },
-    ],
-  });
-  return mentions.map((m) => ({
-    configurationId: m.agentConfigurationId as string,
-  }));
+  return [...globalAgents, ...agents];
 }
 
 /**
@@ -420,7 +283,6 @@ export async function createAgentConfiguration(
   }
 
   let version = 0;
-  let formerAgentRelationOverride: AgentRelationOverrideType | null = null;
 
   const agentConfig = await front_sequelize.transaction(
     async (t): Promise<AgentConfiguration> => {
@@ -438,43 +300,8 @@ export async function createAgentConfiguration(
 
         if (latestVersion !== null) {
           version = latestVersion + 1;
-          const formerAgent = await AgentConfiguration.findOne({
-            where: {
-              sId: agentConfigurationId,
-              workspaceId: owner.id,
-              version: latestVersion,
-            },
-            attributes: ["scope"],
-            include: [
-              {
-                model: AgentUserRelation,
-                where: {
-                  userId: user.id,
-                },
-                attributes: ["relation"],
-              },
-            ],
-            transaction: t,
-          });
-          if (
-            formerAgent?.relationOverrides &&
-            formerAgent.relationOverrides.length > 0
-          ) {
-            formerAgentRelationOverride =
-              formerAgent.relationOverrides[0].relation;
-          }
-          // At time of writing, private agents can only be created from
-          // scratch. An existing agent that is not already private cannot be
-          // updated back to private.
-
-          if (
-            formerAgent &&
-            scope === "private" &&
-            formerAgent.scope !== "private"
-          ) {
-            throw new Error("Published agents cannot go back to private.");
-          }
         }
+
         await AgentConfiguration.update(
           { status: "archived" },
           {
@@ -519,7 +346,6 @@ export async function createAgentConfiguration(
     sId: agentConfig.sId,
     version: agentConfig.version,
     scope: agentConfig.scope,
-    relationOverride: formerAgentRelationOverride,
     name: agentConfig.name,
     description: agentConfig.description,
     pictureUrl: agentConfig.pictureUrl,
