+++ front/lib/api/assistant/global_agents.ts
@@ -100,7 +100,6 @@ async function _getHelperGlobalAgent(
     description: "Help on how to use Dust",
     pictureUrl: "https://dust.tt/static/systemavatar/helper_avatar_full.png",
     status: "active",
-    relationOverride: null,
     scope: "global",
     generation: {
       id: -1,
@@ -127,7 +126,6 @@ async function _getGPT35TurboGlobalAgent({
     pictureUrl: "https://dust.tt/static/systemavatar/gpt3_avatar_full.png",
     status: settings ? settings.status : "active",
     scope: "global",
-    relationOverride: null,
     generation: {
       id: -1,
       prompt: "",
@@ -157,7 +155,6 @@ async function _getGPT4GlobalAgent({
     pictureUrl: "https://dust.tt/static/systemavatar/gpt4_avatar_full.png",
     status,
     scope: "global",
-    relationOverride: null,
     generation: {
       id: -1,
       prompt: "",
@@ -187,7 +184,6 @@ async function _getClaudeInstantGlobalAgent({
     pictureUrl: "https://dust.tt/static/systemavatar/claude_avatar_full.png",
     status: settings ? settings.status : "active",
     scope: "global",
-    relationOverride: null,
     generation: {
       id: -1,
       prompt: "",
@@ -219,7 +215,6 @@ async function _getClaudeGlobalAgent({
     pictureUrl: "https://dust.tt/static/systemavatar/claude_avatar_full.png",
     status: settings ? settings.status : status,
     scope: "global",
-    relationOverride: null,
     generation: {
       id: -1,
       prompt: "",
@@ -247,7 +242,6 @@ async function _getMistralGlobalAgent({
     pictureUrl: "https://dust.tt/static/systemavatar/mistral_avatar_full.png",
     status: settings ? settings.status : "disabled_by_admin",
     scope: "global",
-    relationOverride: null,
     generation: {
       id: -1,
       prompt: "",
@@ -310,7 +304,6 @@ async function _getManagedDataSourceAgent(
       pictureUrl,
       status: "disabled_by_admin",
       scope: "global",
-      relationOverride: null,
       generation: null,
       action: null,
     };
@@ -330,7 +323,6 @@ async function _getManagedDataSourceAgent(
       pictureUrl,
       status: "disabled_missing_datasource",
       scope: "global",
-      relationOverride: null,
       generation: null,
       action: null,
     };
@@ -345,7 +337,6 @@ async function _getManagedDataSourceAgent(
     pictureUrl,
     status: "active",
     scope: "global",
-    relationOverride: null,
     generation: {
       id: -1,
       prompt,
@@ -487,7 +478,6 @@ async function _getDustGlobalAgent(
       pictureUrl,
       status: "disabled_by_admin",
       scope: "global",
-      relationOverride: null,
       generation: null,
       action: null,
     };
@@ -515,7 +505,6 @@ async function _getDustGlobalAgent(
       pictureUrl,
       status: "disabled_missing_datasource",
       scope: "global",
-      relationOverride: null,
       generation: null,
       action: null,
     };
@@ -530,7 +519,6 @@ async function _getDustGlobalAgent(
     pictureUrl,
     status: "active",
     scope: "global",
-    relationOverride: null,
     generation: {
       id: -1,
       prompt:
