+++ front/lib/models/assistant/agent.ts
@@ -19,10 +19,11 @@ import {
 } from "sequelize";
 
 import { front_sequelize } from "@app/lib/databases";
-import { AgentDustAppRunConfiguration } from "@app/lib/models/assistant/actions/dust_app_run";
 import { AgentRetrievalConfiguration } from "@app/lib/models/assistant/actions/retrieval";
 import { User } from "@app/lib/models/user";
 import { Workspace } from "@app/lib/models/workspace";
+
+import { AgentDustAppRunConfiguration } from "./actions/dust_app_run";
 
 /**
  * Configuration of Agent generation.
@@ -117,7 +118,6 @@ export class AgentConfiguration extends Model<
   declare generationConfiguration: NonAttribute<AgentGenerationConfiguration>;
   declare retrievalConfiguration: NonAttribute<AgentRetrievalConfiguration>;
   declare dustAppRunConfiguration: NonAttribute<DustAppRunConfigurationType>;
-  declare relationOverrides: NonAttribute<AgentUserRelation[]>;
 }
 AgentConfiguration.init(
   {
@@ -368,15 +368,6 @@ Workspace.hasMany(AgentUserRelation, {
 AgentConfiguration.hasMany(AgentUserRelation, {
   foreignKey: { allowNull: false },
   onDelete: "CASCADE",
-});
-AgentUserRelation.belongsTo(User, {
-  foreignKey: { allowNull: false },
-});
-AgentUserRelation.belongsTo(Workspace, {
-  foreignKey: { allowNull: false },
-});
-AgentUserRelation.belongsTo(AgentConfiguration, {
-  foreignKey: { allowNull: false },
 });
 
 AgentUserRelation.addHook(
