+++ front/pages/api/v1/w/[wId]/assistant/agent_configurations.ts
@@ -34,7 +34,7 @@ async function handler(
 
   switch (req.method) {
     case "GET": {
-      const agentConfigurations = await getAgentConfigurations(auth, "all");
+      const agentConfigurations = await getAgentConfigurations(auth);
       return res.status(200).json({
         agentConfigurations,
       });
