+++ front/pages/api/w/[wId]/assistant/agent_configurations/[aId]/index.ts
@@ -1,7 +1,4 @@
-import {
-  AgentConfigurationType,
-  PostOrPatchAgentConfigurationRequestBodySchema,
-} from "@dust-tt/types";
+import { AgentConfigurationType } from "@dust-tt/types";
 import { ReturnedAPIErrorType } from "@dust-tt/types";
 import { isLeft } from "fp-ts/lib/Either";
 import * as reporter from "io-ts-reporters";
@@ -13,7 +10,11 @@ import {
 } from "@app/lib/api/assistant/configuration";
 import { Authenticator, getSession } from "@app/lib/auth";
 import { apiError, withLogging } from "@app/logger/withlogging";
-import { createOrUpgradeAgentConfiguration } from "@app/pages/api/w/[wId]/assistant/agent_configurations";
+
+import {
+  createOrUpgradeAgentConfiguration,
+  PostOrPatchAgentConfigurationRequestBodySchema,
+} from "..";
 
 export type GetAgentConfigurationResponseBody = {
   agentConfiguration: AgentConfigurationType;
