+++ front/pages/w/[wId]/assistant/new.tsx
@@ -97,7 +97,6 @@ export default function AssistantNew({
 
   const { agentConfigurations } = useAgentConfigurations({
     workspaceId: owner.sId,
-    agentsGetViewType: "list",
   });
 
   const activeAgents = agentConfigurations.filter((a) => a.status === "active");
