+++ front/pages/w/[wId]/builder/assistants/dust.tsx
@@ -69,7 +69,6 @@ export default function EditDustAssistant({
   const { agentConfigurations, mutateAgentConfigurations } =
     useAgentConfigurations({
       workspaceId: owner.sId,
-      agentsGetViewType: "all",
     });
   const { dataSources, mutateDataSources } = useDataSources(owner);
