+++ front/pages/w/[wId]/builder/assistants/index.tsx
@@ -70,7 +70,6 @@ export default function AssistantsBuilder({
   const { agentConfigurations, mutateAgentConfigurations } =
     useAgentConfigurations({
       workspaceId: owner.sId,
-      agentsGetViewType: "all",
     });
   const [assistantSearch, setAssistantSearch] = useState<string>("");
   const [showDisabledFreeWorkspacePopup, setShowDisabledFreeWorkspacePopup] =
