+++ types/src/index.ts
@@ -1,19 +1,11 @@
 export * from "./core/data_source";
-export * from "./front/api_handlers/internal/agent_configuration";
-export * from "./front/api_handlers/internal/assistant";
 export * from "./front/api_handlers/internal/assistant";
 export * from "./front/api_handlers/public/assistant";
-export * from "./front/api_handlers/public/assistant";
-export * from "./front/api_handlers/public/data_sources";
 export * from "./front/api_handlers/public/data_sources";
 export * from "./front/app";
 export * from "./front/assistant/actions/dust_app_run";
-export * from "./front/assistant/actions/dust_app_run";
-export * from "./front/assistant/actions/retrieval";
 export * from "./front/assistant/actions/retrieval";
 export * from "./front/assistant/agent";
-export * from "./front/assistant/agent";
-export * from "./front/assistant/conversation";
 export * from "./front/assistant/conversation";
 export * from "./front/data_source";
 export * from "./front/dataset";
@@ -38,14 +30,9 @@ export * from "./front/lib/error";
 export * from "./front/lib/extract_event_app";
 export * from "./front/lib/result";
 export * from "./front/membership_invitation";
-export * from "./front/membership_invitation";
-export * from "./front/plan";
 export * from "./front/plan";
 export * from "./front/project";
-export * from "./front/project";
 export * from "./front/provider";
-export * from "./front/provider";
-export * from "./front/run";
 export * from "./front/run";
 export * from "./front/user";
 export * from "./shared/cache";
